Move Util code dealing with resources out of libanki
