+++ AnkiDroid/src/main/java/com/ichi2/anki/Reviewer.kt
@@ -73,6 +73,7 @@ import com.ichi2.anki.servicelayer.NoteService.toggleMark
 import com.ichi2.anki.servicelayer.rescheduleCards
 import com.ichi2.anki.servicelayer.resetCards
 import com.ichi2.anki.snackbar.showSnackbar
+import com.ichi2.anki.utils.remainingTime
 import com.ichi2.anki.workarounds.FirefoxSnackbarWorkaround.handledLaunchFromWebBrowser
 import com.ichi2.annotations.NeedsTest
 import com.ichi2.libanki.*
@@ -948,7 +949,7 @@ open class Reviewer :
             if (mPrefShowETA) {
                 launchCatchingTask {
                     mEta = withCol { sched.eta(counts, false) }
-                    actionBar.subtitle = Utils.remainingTime(this@Reviewer, (mEta * 60).toLong())
+                    actionBar.subtitle = remainingTime(this@Reviewer, (mEta * 60).toLong())
                 }
             }
         }
